Move pomodoro-status file location
